10203 - Fix flaky preview pane script injection
Use the document passed to the preview component rather than manually searching for it from the outside.
