+++ static/admin/index.html
@@ -10,7 +10,6 @@
     <script defer src="react-dom.production.min.js"></script>
     <script defer src="decap-cms-app.js"></script>
     <script defer src="cms.js"></script>
-    <script defer src="inject-preview-pane-dependencies.js"></script>
   </head>
   <body>
   </body>
